+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".MapsActivity"
+            android:name=".ui.MapsActivity"
             android:label="@string/title_activity_maps">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -40,15 +40,15 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".TUC_Info"
+            android:name=".ui.InfoActivity"
             android:label="TUCInfo">
         </activity>
         <activity
-            android:name=".TUC_Photos"
+            android:name=".ui.PhotosActivity"
             android:label="Photos">
         </activity>
         <activity
-            android:name=".CECH_Info"
+            android:name=".ui.CECHInfoActivity"
             android:label="CECHInfo">
         </activity>
     </application>
