--- app/src/main/java/com/example/tour_a_versity_v3/TUC_Photos.kt
@@ -1,11 +1,0 @@
-package com.example.tour_a_versity_v3
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class TUC_Photos : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity3)
-    }
-}
