+++ app/src/main/java/com/example/tour_a_versity_v3/ui/InfoActivity.kt
@@ -1,18 +1,19 @@
-package com.example.tour_a_versity_v3
+package com.example.tour_a_versity_v3.ui
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
+import com.example.tour_a_versity_v3.R
 
-class TUC_Info : AppCompatActivity() {
+class InfoActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity2)
 
         val button = findViewById<Button>(R.id.TucPhotoButton)
         button.setOnClickListener {
-            val intent = Intent(this, TUC_Photos::class.java)
+            val intent = Intent(this, PhotosActivity::class.java)
             startActivity(intent)
         }
     }
