+++ app/src/main/java/com/example/tour_a_versity_v3/ui/PhotosActivity.kt
@@ -1,10 +1,10 @@
-package com.example.tour_a_versity_v3
+package com.example.tour_a_versity_v3.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.tour_a_versity_v3.R
 
-class CECH_Photos: AppCompatActivity() {
+class PhotosActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity3)
